infra/requirements.txt
aws-cdk-lib==2.100.0
constructs>=10.0.0,<11.0.0
pytest==7.4.3
pytest-cov==4.1.0